refactor : Permission 요청 함수 분리 및, SettingPreferenceManager 오브젝트 패키지 이동
