+++ app/src/main/java/woowacourse/movie/SettingPreferenceManager.kt
@@ -1,12 +1,11 @@
-package woowacourse
+package woowacourse.movie
 
 import android.content.Context
 import android.content.SharedPreferences
 import androidx.preference.PreferenceManager
 
-private const val KEY_PUSH_ALARM = "KEY_PUSH_ALARM"
-
 object SettingPreferenceManager {
+    private const val KEY_PUSH_ALARM = "KEY_PUSH_ALARM"
     private lateinit var sharedPreference: SharedPreferences
     fun inIt(context: Context) {
         sharedPreference = PreferenceManager.getDefaultSharedPreferences(context)
