+++ app/src/main/res/layout/item_reservation.xml
@@ -3,7 +3,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
-        android:layout_height="66dp"
+        android:layout_height="wrap_content"
         android:background="?attr/selectableItemBackground">
 
     <TextView
@@ -22,12 +22,22 @@
             android:id="@+id/item_reservation_title"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginBottom="12dp"
+            android:paddingBottom="12dp"
             android:textColor="@color/black"
             android:textSize="16sp"
             android:textStyle="bold"
-            app:layout_constraintBottom_toBottomOf="parent"
+            android:layout_marginTop="15dp"
+            app:layout_constraintTop_toBottomOf="@id/item_reservation_date_time"
             app:layout_constraintStart_toStartOf="@+id/item_reservation_date_time"
             tools:text="해리 포터와 마법사의 돌" />
+
+    <View
+            android:layout_width="0dp"
+            android:layout_height="1dp"
+            android:background="@color/black"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintBottom_toBottomOf="parent"
+            tools:visibility="visible" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
