+++ app/src/main/res/values/strings.xml
@@ -37,4 +37,5 @@
     <string name="reservation_date_time_format">"yyyy-MM-dd | HH:mm"</string>
     <string name="setting_title">푸시 알람 수신</string>
     <string name="setting_push_alarm_description">해제하면 푸시 알림을 수신할 수 없습니다.</string>
+    <string name="notification_refuse_two_time_warning_message">만약 이번에도 권한을 거부한다면, 다시 앱을 설치하셔야 권한 허용이 가능합니다</string>
 </resources>
